# openapi_copilot.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PO Tools API",
    "description": "API for Agentic interaction with PO Tools",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/generate-issue": {
      "post": {
        "tags": [
          "AI Issue Creator"
        ],
        "summary": "Generate Issue",
        "operationId": "generate_issue_api_v1_generate_issue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/create-jira-issue": {
      "post": {
        "tags": [
          "AI Issue Creator"
        ],
        "summary": "Create Jira Issue Endpoint",
        "operationId": "create_jira_issue_endpoint_api_v1_create_jira_issue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJiraIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateJiraIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/kpis/calculate": {
      "post": {
        "tags": [
          "Sprint KPIs"
        ],
        "summary": "Calculate Kpis",
        "operationId": "calculate_kpis_api_v1_kpis_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculateKPIsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/kpis/publish": {
      "post": {
        "tags": [
          "Sprint KPIs"
        ],
        "summary": "Publish Kpis",
        "operationId": "publish_kpis_api_v1_kpis_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishKPIsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/release-notes/prepare": {
      "post": {
        "tags": [
          "Release Notes"
        ],
        "summary": "Prepare Release Notes",
        "operationId": "prepare_release_notes_api_v1_release_notes_prepare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareReleaseNotesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/release-notes/publish": {
      "post": {
        "tags": [
          "Release Notes"
        ],
        "summary": "Publish Release Notes",
        "operationId": "publish_release_notes_api_v1_release_notes_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishReleaseNotesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/release-notes/generate-pdf": {
      "post": {
        "tags": [
          "Release Notes"
        ],
        "summary": "Generate Release Notes Pdf",
        "operationId": "generate_release_notes_pdf_api_v1_release_notes_generate_pdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReleaseNotesPdfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/sprint-review/generate-pdf": {
      "post": {
        "tags": [
          "Sprint Review"
        ],
        "summary": "Generate Sprint Review Pdf",
        "operationId": "generate_sprint_review_pdf_api_v1_sprint_review_generate_pdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSprintReviewPdfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/sprint-review/publish": {
      "post": {
        "tags": [
          "Sprint Review"
        ],
        "summary": "Publish Sprint Review",
        "operationId": "publish_sprint_review_api_v1_sprint_review_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishSprintReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/quarterly-planner/generate-roadmap-pdf": {
      "post": {
        "tags": [
          "Quarterly Planner"
        ],
        "summary": "Generate Roadmap Pdf",
        "operationId": "generate_roadmap_pdf_api_v1_quarterly_planner_generate_roadmap_pdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRoadmapPdfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/quarterly-planner/generate-progress-pdf": {
      "post": {
        "tags": [
          "Quarterly Planner"
        ],
        "summary": "Generate Progress Pdf",
        "operationId": "generate_progress_pdf_api_v1_quarterly_planner_generate_progress_pdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateProgressPdfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateJiraIssueResponse": {
        "properties": {
          "issue_key": {
            "type": "string",
            "title": "Issue Key"
          },
          "issue_link": {
            "type": "string",
            "title": "Issue Link"
          }
        },
        "type": "object",
        "required": [
          "issue_key",
          "issue_link"
        ],
        "title": "CreateJiraIssueResponse"
      },
      "GenerateIssueRequest": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "issue_type": {
            "type": "string",
            "title": "Issue Type"
          },
          "input_text": {
            "type": "string",
            "title": "Input Text"
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "issue_type",
          "input_text"
        ],
        "title": "GenerateIssueRequest"
      },
      "GenerateIssueResponse": {
        "properties": {
          "generated_description": {
            "type": "string",
            "title": "Generated Description"
          }
        },
        "type": "object",
        "required": [
          "generated_description"
        ],
        "title": "GenerateIssueResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "servers": [
    {
      "url": "https://TU-TUNEL.trycloudflare.com",
      "description": "Cloudflare Tunnel"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    }
  ]
}